pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chuntaro-chats-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.1",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.1",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlmigrate>=1.0.0",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]
